Ensure setter 'value' is preselected'
